fix: set minimum go version to 1.12, allow broader integration
The `go` directive in go.mod sets the _minimum_ required version of go
for the module to work. Previously, this was set to `1.25.3`, which is a
very recent version of go. This means that in order to use this library,
all downstream projects must also bump their build systems to use `go
1.25.3+`.

Using the tool [mingo](https://github.com/bobg/mingo), it's easy to
validate that the actual functionality of the package only needs code
within `go 1.12+`:

```
~/go/src/github.com/alpkeskin/gotoon (main [ ]) -> mingo -v
/home/tjhop/go/src/github.com/alpkeskin/gotoon/encoders.go:36:3: 1 (function body with no final return statement)
/home/tjhop/go/src/github.com/alpkeskin/gotoon/primitives.go:55:10: 12 ("strings".ReplaceAll)
12
```

By reducing the required go version here, it will make it easier and
less disruptive to adopt the library downstream.

Signed-off-by: TJ Hoplock <t.hoplock@gmail.com>